security(engines): update template
This PR addresses multiple security issues in npm and yarn that is using latest known "safe" versions. Further, node 8 has reached end of life and is no longer receiving security backports. It should be noted that some CI might not be able to process yarn above 1.6 - however that can be changed easily by a savvy developer.