chore(security): update engines object in template's package.json (#83)
* * .gitignore: add /src-cordova/node_modules that is generated by Cordova 8+ (#46)

* ESLint ignore paths beginning with . (#47)

* fix: Temporary until ajv-errors package gets a proper fix

* Revert "fix: Temporary until ajv-errors package gets a proper fix"

This reverts commit 955c710778de19f75b6bb8d910fdad93d5b8e6fe.

* chore: Update README

* Update to include upcoming BEX mode. (#70)

We don't need the dev files (www) or the core files which are copied across on each build (js/core)

* chore(bex): Include chrome in ignores for BEX mode (#71)

`chrome` is used to interact with the extension.

* security(engines): update template

This PR addresses multiple security issues in npm and yarn that is using latest known "safe" versions. Further, node 8 has reached end of life and is no longer receiving security backports. It should be noted that some CI might not be able to process yarn above 1.6 - however that can be changed easily by a savvy developer.

Co-authored-by: Ciprian Popa <ciprian.popa@jaguard.com>
Co-authored-by: fansanelli <francians@gmail.com>
Co-authored-by: Razvan Stoenescu <razvan.stoenescu@gmail.com>
Co-authored-by: webnoob <allangaunt@gmail.com>